"Create Golang Slack bot for calculating age"
Description:

Developed a Slack bot using Golang that enables users to calculate their age based on their year of birth.
Features:

Users can mention the bot and provide their year of birth to receive their age as a response in Slack.
Implemented event subscriptions for the bot to handle app mentions and relevant chat events.
Set up the necessary environment variables for Slack API integration.
Prerequisites:

Installed Go language (https://golang.org/dl/).
Configured Slack workspace with administrative access.
Instructions:

Replace the placeholders for Slack Bot Token and Socket Token in the main.go file with the actual tokens from the Slack API.
Obtain the required tokens by creating a new app in api.slack.com/apps and enabling socket mode and event subscriptions.
Install the bot to the workspace and obtain the Bot User OAuth Token for socket mode.
To Run the Bot:

Initialize the project folder with go mod init github.com/theafwan/slack-age-bot.
Install the required package with go get "github.com/shomali11/slacker" and tidy the dependencies with go mod tidy.
Build the project using go build and execute it with go run main.go.
In Slack, add the bot to the desired channel and mention it with the year of birth to get the age as a response.